--- helpers/fuzzystrmatcher/blacklist.go
@@ -1,200 +1,0 @@
-package fuzzystrmatcher
-
-var allBlackListedWords = append(blackListedEnglishWords, blackListedDutchWords...)
-
-// Common english words that not not really add any value when searching
-var blackListedEnglishWords = []string{
-	"the",
-	"be",
-	"to",
-	"of",
-	"and",
-	"a",
-	"in",
-	"that",
-	"have",
-	"i",
-	"it",
-	"for",
-	"not",
-	"on",
-	"he",
-	"as",
-	"you",
-	"do",
-	"at",
-	"this",
-	"but",
-	"his",
-	"by",
-	"from",
-	"they",
-	"we",
-	"say",
-	"her",
-	"she",
-	"or",
-	"an",
-	"will",
-	"my",
-	"one",
-	"all",
-	"would",
-	"there",
-	"their",
-	"what",
-	"so",
-	"up",
-	"out",
-	"if",
-	"about",
-	"who",
-	"get",
-	"which",
-	"go",
-	"me",
-	"when",
-	"make",
-	"can",
-	"like",
-	"time",
-	"no",
-	"just",
-	"him",
-	"know",
-	"take",
-	"people",
-	"into",
-	"year",
-	"your",
-	"good",
-	"some",
-	"could",
-	"them",
-	"see",
-	"other",
-	"than",
-	"then",
-	"now",
-	"look",
-	"only",
-	"come",
-	"its",
-	"over",
-	"think",
-	"also",
-	"back",
-	"after",
-	"use",
-	"two",
-	"how",
-	"our",
-	"work",
-	"first",
-	"well",
-	"way",
-	"even",
-	"new",
-	"want",
-	"because",
-	"any",
-	"these",
-	"give",
-	"day",
-	"most",
-	"us",
-	"are",
-}
-
-// Common dutch word that do not really add any value when searching
-var blackListedDutchWords = []string{
-	"ik",
-	"je",
-	"het",
-	"de",
-	"is",
-	"dat",
-	"een",
-	"niet",
-	"en",
-	"wat",
-	"van",
-	"we",
-	"in",
-	"ze",
-	"op",
-	"te",
-	"hij",
-	"zijn",
-	"er",
-	"maar",
-	"me",
-	"die",
-	"heb",
-	"voor",
-	"met",
-	"als",
-	"ben",
-	"was",
-	"mijn",
-	"u",
-	"dit",
-	"aan",
-	"om",
-	"hier",
-	"naar",
-	"dan",
-	"jij",
-	"zo",
-	"weet",
-	"ja",
-	"kan",
-	"geen",
-	"nog",
-	"wel",
-	"wil",
-	"moet",
-	"goed",
-	"hem",
-	"hebben",
-	"nee",
-	"heeft",
-	"waar",
-	"nu",
-	"hoe",
-	"ga",
-	"kom",
-	"uit",
-	"gaan",
-	"bent",
-	"haar",
-	"doen",
-	"ook",
-	"mij",
-	"over",
-	"of",
-	"daar",
-	"zou",
-	"al",
-	"jullie",
-	"zal",
-	"bij",
-	"ons",
-	"gaat",
-	"hebt",
-	"meer",
-	"waarom",
-	"iets",
-	"deze",
-	"laat",
-	"had",
-	"doe",
-	"moeten",
-	"wie",
-	"jou",
-	"alles",
-	"denk",
-	"kunnen",
-	"eens",
-	"echt",
-}
